Add markers of interest to level response trajectories

"Markers" are used to specify events of interest along response trajectories across individual lanes.

add_marker(data = NULL, mapping = aes(), ...)

Arguments

data

a dataframe prepared for use with ggswim(), either coming from a parent ggswim() function, another add_marker() call, or a new dataframe prepared for use with ggswim().

mapping

Set of aesthetic mappings created by aes(). If specified and inherit.aes = TRUE (the default), it is combined with the default mapping at the top level of the plot. You must supply mapping if there is no plot mapping. More information about accepted mapping arguments can be found in Aesthetics.

...

Other arguments passed to add_marker(), often aesthetic fixed values, i.e. color = "red" or size = 3.

Value

A ggswim object

Aesthetics

add_marker() understands the following aesthetics (required aesthetics are in bold) when using color/colour akin to geom_point()

x

y

color/colour

alpha

group

shape

size

stroke

add_marker() understands the following aesthetics (required aesthetics are in bold) when using label akin to geom_label(). See "Notes" below for additional considerations and requirements.

x

y

color/colour

label *

name *

alpha

angle

family

fontface

group

hjust

lineheight

size

vjust

Note: add_marker() does not support mapping using fill.

Note: If using label, a secondary color reference is required to make the label appear in the color layer of the legend.

Note: If using a dynamic color call, a specifying name is required for aesthetic mapping to render the legend correctly.