check for missing params

Utility check to block users from submitting functions that are missing key parameters. Tests a single parameter for being NULL.

Usage

check_missing_params(param, name, parent_func)

Arguments

param

A parameter to check for NULL

name

The name of the parameter checked

parent_func

The function in which this is being called, to be referenced in the message output