Reshape lane_colors

In order to assign the correct colors to ggplot2::scale_color_manual(), it is necessary to account for the order of the legend output in combination with the assigned markers.

get_assigned_line_colors(gg, lane_colors)

Arguments

gg

a ggplot object

lane_colors

a vector of character strings to assign to the lanes, written in the same order. Optional, if left NULL, default colors applied

Value

a vector