Create mapping object

To assist with label legend fixing, the mapping object helps pass mapped data from aes() down to get_label_override(), including in instances when data is inherited from ggplot().

Usage

get_mapping_obj(object_mapping, plot_mapping)

Arguments

object_mapping

mapping from the layer object

plot_mapping

mapping from the plot object

Value

a aes() frame