Plot swimmer data

Plot swimmer data.

ggswim(
  df,
  id,
  time,
  events,
  reference_event,
  markers,
  lanes,
  lane_colors = NULL,
  title = NULL,
  xlab = NULL,
  ylab = NULL
)

Arguments

df

a dataframe prepared for use with ggswim()

id

the y-axis variable of a swimmer plot, typically a unique subject or record identification column

time

the x-axis variable of the swimmer plot, typically a function of time

events

the column that will supply definitions for the reference_event, markers, and lanes

reference_event

a character string found in events that establishes the time-zero reference point for the x-axis

markers

A character vector that will comprise point markers on the swimmer plot. Optional, default NULL

lanes

Columns that indicate line changes, i.e. color changes for individual swim lanes.

lane_colors

a vector of character strings to assign to the lanes, written in the same order. Optional, if left NULL, default colors applied

title

the plot title

xlab

the x-axis label

ylab

the y-axis label

Value

a ggplot2 figure