Patient Data dataset

These datasets provide information about patients, infusion events, and end of study events. Trial data has been de-identified and randomized for general use.

Usage

data(patient_data)

Format

An object of class tbl_df (inherits from tbl, data.frame) with 53 rows and 4 columns.

Source

This dataset is for demonstration purposes only.

Details

patient_data contains pre-formatted time series data related to disease status markers and status markers that can support arrows. This dataset is most applicable for use with ggswim() and add_marker().

Examples

patient_data
#> # A tibble: 53 × 4
#>    pt_id disease_assessment       start_time end_time
#>    <chr> <chr>                         <dbl>    <dbl>
#>  1 01    CR/CRi + B Cell Recovery       -2.8      0  
#>  2 01    RD                              0        0.9
#>  3 01    CR/CRi + B Cell Aplasia         0.9      1.8
#>  4 01    CR/CRi + B Cell Recovery        1.8      2  
#>  5 02    RD                             -2.8      0  
#>  6 02    CRi                             0        0.2
#>  7 03    CR/CRi + B Cell Recovery       -2.4      0  
#>  8 03    CR/CRi + B Cell Recovery        0        0.9
#>  9 03    CR/CRi + B Cell Aplasia         0.9      2.8
#> 10 03    CR/CRi + B Cell Aplasia         2.8      6.2
#> # ℹ 43 more rows