Retrieve original vars from coerced vars

Detect instances where users manipulate aes() names that ggswim requires to access and identify layer types downstream. In instances where such a coercion is detected, attempt to retrieve the original aes() name.

Usage

retrieve_original_aes(data, aes_mapping, aes_var)

Arguments

data

the data responsible for the current layer

aes_mapping

a list of mapping data (i.e. unlist(mapping))

aes_var

the aesthetic variable to test for (ex: color, shape)

Value

The original variable name as a character string

Details

ggswim references internal ggplot layers and any aesthetic mapping required for downstream rendering. If a user applies a coercion in a function, ggswim may not be able to parse the original variable. For example, in ggswim(mtcars, aes(x = hp, y = cyl, color = factor(disp)), rlang::get_expr() will see the color mapping aesthetic as factor(disp), and not disp.