Add markers of interest to level response trajectories

"Markers" are used to specify events of interest along response trajectories across individual lanes. add_marker() wraps geom_point() or geom_label() depending on the users designation of color/colour and label arguments.

See Aesthetics for more information.

Usage

add_marker(mapping = aes(), data = NULL, ...)

Arguments

mapping

Set of aesthetic mappings created by aes(). If specified and inherit.aes = TRUE (the default), it is combined with the default mapping at the top level of the plot. You must supply mapping if there is no plot mapping.

data

a dataframe prepared for use with ggswim(), either coming from a parent ggswim() function, another add_marker() call, or a new dataframe prepared for use with ggswim().

...

Other arguments passed on to layer(). These are often aesthetics, used to set an aesthetic to a fixed value, like colour = "red" or size = 3. They may also be parameters to the paired geom/stat.

Value

A ggswim object

Aesthetics

add_marker() understands the following aesthetics (required aesthetics are in bold) when using color/colour similar to geom_point()

x

y

color/colour

name *

alpha

group

shape

size

stroke

add_marker() understands the following aesthetics (required aesthetics are in bold) when using label similar to geom_label(). See "Notes" below for additional considerations and requirements.

x

y

label_vals

label_names *

alpha

angle

family

fontface

group

hjust

lineheight

size

vjust

Notes:

add_marker() does not support mapping using fill.

If using a static/non-mapping color specifier, a mapping name is required for aesthetic mapping to render the legend correctly.

If using labels, both label_vals and label_names are required for proper legend population. At minimum, label_vals is needed for data display. These are unique parameter options for aes() to ggswim.

Examples

# markers with points and aesthetic mapping params
add_marker(
  data = infusion_events,
  mapping = aes(
    x = delta_t0,
    y = pt_id,
    color = infusion_type,
    shape = end_study_name
  ),
  size = 5
)
#> mapping: x = ~delta_t0, y = ~pt_id, colour = ~infusion_type, shape = ~end_study_name 
#> geom_point: na.rm = TRUE
#> stat_identity: na.rm = TRUE
#> position_identity 

# markers with points and static params

initial_infusions <- infusion_events[infusion_events$infusion_type == "Infusion", ]

add_marker(
  data = initial_infusions,
  mapping = aes(
    x = delta_t0,
    y = pt_id,
    name = "Initial"
  ),
  color = "red",
  size = 5
)
#> Warning: Duplicated aesthetics after name standardisation: colour
#> mapping: x = ~delta_t0, y = ~pt_id, colour = ~Initial 
#> geom_point: na.rm = TRUE
#> stat_identity: na.rm = TRUE
#> position_identity 

# markers with labels
add_marker(
  data = end_study_events,
  mapping = aes(y = pt_id, x = delta_t0_months,
      label_names = end_study_name,
      label_vals = end_study_label
  ),
  label.size = NA, fill = NA, size = 5
)
#> mapping: x = ~delta_t0_months, y = ~pt_id, colour = ~end_study_name, label = ~end_study_label 
#> geom_label: parse = FALSE, label.padding = 0.25, label.r = 0.15, label.size = NA, size.unit = mm, na.rm = FALSE
#> stat_identity: na.rm = FALSE
#> position_identity