Add arrows to swimmer plot lanes

Arrows attached to the end of swimmer plot lanes can be used to denote the continuation of events such as ongoing treatment, implying that the activity or status extends beyond the plotted period.

Usage

geom_swim_arrow(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  ...,
  arrow_colour = "black",
  arrow_head_length = unit(0.25, "inches"),
  arrow_neck_length = NULL,
  arrow_fill = NULL,
  arrow_type = "closed",
  lineend = "butt",
  linejoin = "round",
  na.rm = FALSE,
  show.legend = FALSE,
  inherit.aes = TRUE
)

Arguments

mapping

Set of aesthetic mappings created by aes(). If specified and inherit.aes = TRUE (the default), it is combined with the default mapping at the top level of the plot. You must supply mapping if there is no plot mapping.

data

A dataframe prepared for use with geom_swim_arrow()

stat

The statistical transformation to use on the data for this layer. When using a geom_*() function to construct a layer, the stat argument can be used the override the default coupling between geoms and stats. The stat argument accepts the following:

A Stat ggproto subclass, for example StatCount.

A string naming the stat. To give the stat as a string, strip the function name of the stat_ prefix. For example, to use stat_count(), give the stat as "count".

For more information and other ways to specify the stat, see the layer stat documentation.

position

Position adjustment. ggswim accepts either "stack", or "identity" depending on the use case. Default "identity".

...

Other arguments passed on to layer()'s params argument. These arguments broadly fall into one of 4 categories below. Notably, further arguments to the position argument, or aesthetics that are required can not be passed through .... Unknown arguments that are not part of the 4 categories below are ignored.

Static aesthetics that are not mapped to a scale, but are at a fixed value and apply to the layer as a whole. For example, colour = "red" or linewidth = 3. The geom's documentation has an Aesthetics section that lists the available options. The 'required' aesthetics cannot be passed on to the params. Please note that while passing unmapped aesthetics as vectors is technically possible, the order and required length is not guaranteed to be parallel to the input data.

When constructing a layer using a stat_*() function, the ... argument can be used to pass on parameters to the geom part of the layer. An example of this is stat_density(geom = "area", outline.type = "both"). The geom's documentation lists which parameters it can accept.

Inversely, when constructing a layer using a geom_*() function, the ... argument can be used to pass on parameters to the stat part of the layer. An example of this is geom_area(stat = "density", adjust = 0.5). The stat's documentation lists which parameters it can accept.

The key_glyph argument of layer() may also be passed on through .... This can be one of the functions described as key glyphs, to change the display of the layer in the legend.

arrow_colour

The color of the arrow head

arrow_head_length

A unit specifying the length of the arrow head (from tip to base).

arrow_neck_length

Value specifying neck length from end of segment to arrow head base

arrow_fill

The fill color of the arrow head

arrow_type

One of "open" or "closed" indicating whether the arrow head should be a closed triangle.

lineend

Line end style (round, butt, square).

linejoin

Line join style (round, mitre, bevel).

na.rm

If FALSE, the default, missing values are removed with a warning. If TRUE, missing values are silently removed.

show.legend

logical. Should this layer be included in the legends? NA, the default, includes if any aesthetics are mapped. FALSE never includes, and TRUE always includes. It can also be a named logical vector to finely select the aesthetics to display.

inherit.aes

If FALSE, overrides the default aesthetics, rather than combining with them. This is most useful for helper functions that define both data and aesthetics and shouldn't inherit behaviour from the default plot specification, e.g. borders().

Aesthetics

geom_swim_arrow() understands the following aesthetics (required aesthetics are in bold):

x

y

xend

alpha

colour

group

linetype

linewidth

geom_swim_arrow() is a wrapper for geom_segment() and can support much of the same functionality.

Examples

# Set up data for arrows
arrow_data <- patient_data |>
  dplyr::left_join(
    end_study_events |>
      dplyr::select(pt_id, end_study_name),
    by = "pt_id"
  ) |>
  dplyr::select(pt_id, end_time, end_study_name) |>
  dplyr::filter(.by = pt_id, end_time == max(end_time)) |>
  unique()

geom_swim_arrow(
  data = arrow_data,
  mapping = aes(xend = end_time, y = pt_id),
  linewidth = .1,
  arrow_neck_length = 5,
  arrow_head_length = grid::unit(0.25, "inches"),
  arrow_colour = "slateblue",
  arrow_fill = "cyan"
)
#> mapping y = ~pt_id, xend = ~end_time 
#> geom_swim_arrow: na.rm = FALSE, lineend = butt, linejoin = round, arrow = NULL, arrow.fill = cyan, linewidth = 0.1 
#> stat_identity: na.rm = FALSE 
#> position_identity