Patient Dataset

This dataset provides information about patients, their status, adverse events, and medication administration.

Usage

data(patient_data)

patient_status

adverse_events

medication_administration

Format

A list containing three tibbles:

patient_status

A tibble containing information about patient status including subject ID, cohort, status, name, value, and time sorting.

adverse_events

A tibble containing information about adverse events including subject ID, adverse event name, and time of event.

medication_administration

A tibble containing information about medication administration including subject ID, medication name, name, and time of event.

An object of class tbl_df (inherits from tbl, data.frame) with 30 rows and 6 columns.

An object of class tbl_df (inherits from tbl, data.frame) with 14 rows and 3 columns.

An object of class tbl_df (inherits from tbl, data.frame) with 8 rows and 4 columns.

Source

This dataset is for demonstration purposes only.

Examples

patient_data
#> $patient_status
#> # A tibble: 30 × 6
#>    subject_id cohort   status    name       value time_sorting
#>    <fct>      <fct>    <ord>     <chr>      <dbl>        <dbl>
#>  1 9          Cohort B On Study  time_start  0            4.98
#>  2 9          Cohort B On Study  time_end    4.98         4.98
#>  3 10         Cohort B On Study  time_start  0            8.19
#>  4 10         Cohort B On Study  time_end    8.19         8.19
#>  5 4          Cohort B On Study  time_start  0           24.2 
#>  6 4          Cohort B On Study  time_end    4.73        24.2 
#>  7 4          Cohort B Off Study time_start  4.73        24.2 
#>  8 4          Cohort B Off Study time_end   14.7         24.2 
#>  9 8          Cohort A On Study  time_start  0            1.41
#> 10 8          Cohort A On Study  time_end    1.41         1.41
#> # ℹ 20 more rows
#> 
#> $adverse_events
#> # A tibble: 14 × 3
#>    subject_id adverse_event_name   time_of_event
#>    <fct>      <chr>                        <dbl>
#>  1 1          Psychiatric Disorder         1.79 
#>  2 1          Psychiatric Disorder        11.0  
#>  3 1          Infection                    1.67 
#>  4 1          Infection                   19.5  
#>  5 2          Cardiac Disorder             3.10 
#>  6 2          Cardiac Disorder            10.5  
#>  7 3          Psychiatric Disorder         2.76 
#>  8 3          Psychiatric Disorder         5.26 
#>  9 3          Infection                    1.92 
#> 10 3          Infection                    4.32 
#> 11 4          Cardiac Disorder             1.26 
#> 12 4          Cardiac Disorder             8.59 
#> 13 7          Psychiatric Disorder         0.117
#> 14 9          Infection                    1.90 
#> 
#> $medication_administration
#> # A tibble: 8 × 4
#>   subject_id medication name         time_of_event
#>   <fct>      <chr>      <chr>                <dbl>
#> 1 1          ❌         Medication B         1.54 
#> 2 1          ❌         Medication B        13.2  
#> 3 2          ✅         Medication A         0.504
#> 4 2          ✅         Medication A         8.79 
#> 5 4          ❌         Medication B         0.835
#> 6 4          ❌         Medication B        11.6  
#> 7 7          ✅         Medication A         3.35 
#> 8 9          ❌         Medication B         3.83 
#> 

patient_data$patient_status
#> # A tibble: 30 × 6
#>    subject_id cohort   status    name       value time_sorting
#>    <fct>      <fct>    <ord>     <chr>      <dbl>        <dbl>
#>  1 9          Cohort B On Study  time_start  0            4.98
#>  2 9          Cohort B On Study  time_end    4.98         4.98
#>  3 10         Cohort B On Study  time_start  0            8.19
#>  4 10         Cohort B On Study  time_end    8.19         8.19
#>  5 4          Cohort B On Study  time_start  0           24.2 
#>  6 4          Cohort B On Study  time_end    4.73        24.2 
#>  7 4          Cohort B Off Study time_start  4.73        24.2 
#>  8 4          Cohort B Off Study time_end   14.7         24.2 
#>  9 8          Cohort A On Study  time_start  0            1.41
#> 10 8          Cohort A On Study  time_end    1.41         1.41
#> # ℹ 20 more rows

patient_status
#> # A tibble: 30 × 6
#>    subject_id cohort   status    name       value time_sorting
#>    <fct>      <fct>    <ord>     <chr>      <dbl>        <dbl>
#>  1 9          Cohort B On Study  time_start  0            4.98
#>  2 9          Cohort B On Study  time_end    4.98         4.98
#>  3 10         Cohort B On Study  time_start  0            8.19
#>  4 10         Cohort B On Study  time_end    8.19         8.19
#>  5 4          Cohort B On Study  time_start  0           24.2 
#>  6 4          Cohort B On Study  time_end    4.73        24.2 
#>  7 4          Cohort B Off Study time_start  4.73        24.2 
#>  8 4          Cohort B Off Study time_end   14.7         24.2 
#>  9 8          Cohort A On Study  time_start  0            1.41
#> 10 8          Cohort A On Study  time_end    1.41         1.41
#> # ℹ 20 more rows

adverse_events
#> # A tibble: 14 × 3
#>    subject_id adverse_event_name   time_of_event
#>    <fct>      <chr>                        <dbl>
#>  1 1          Psychiatric Disorder         1.79 
#>  2 1          Psychiatric Disorder        11.0  
#>  3 1          Infection                    1.67 
#>  4 1          Infection                   19.5  
#>  5 2          Cardiac Disorder             3.10 
#>  6 2          Cardiac Disorder            10.5  
#>  7 3          Psychiatric Disorder         2.76 
#>  8 3          Psychiatric Disorder         5.26 
#>  9 3          Infection                    1.92 
#> 10 3          Infection                    4.32 
#> 11 4          Cardiac Disorder             1.26 
#> 12 4          Cardiac Disorder             8.59 
#> 13 7          Psychiatric Disorder         0.117
#> 14 9          Infection                    1.90 

medication_administration
#> # A tibble: 8 × 4
#>   subject_id medication name         time_of_event
#>   <fct>      <chr>      <chr>                <dbl>
#> 1 1          ❌         Medication B         1.54 
#> 2 1          ❌         Medication B        13.2  
#> 3 2          ✅         Medication A         0.504
#> 4 2          ✅         Medication A         8.79 
#> 5 4          ❌         Medication B         0.835
#> 6 4          ❌         Medication B        11.6  
#> 7 7          ✅         Medication A         3.35 
#> 8 9          ❌         Medication B         3.83